Change product's type in SavedProduct entity

+++ src/main/java/com/cheaclo/userdatabase/entity/SavedProduct.java
@@ -25,5 +25,5 @@ public class SavedProduct {
     @ManyToOne
     @JoinColumn(name = "user_id")
     private User user;
-    private int product;
+    private Long product;
 }
